# swagger-schema.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/customers/create": {
      "patch": {
        "operationId": "CustomerController_createCustomer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Customer created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "409": {
            "description": "Customer with this platformId, email, CPF, or CNPJ already exists"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new customer with tags",
        "tags": [
          "Customers"
        ]
      }
    },
    "/customers/list": {
      "get": {
        "operationId": "CustomerController_findAllCustomers",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 10,
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term",
            "schema": {
              "example": "john",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by status",
            "schema": {
              "example": "active",
              "type": "string",
              "enum": [
                "active",
                "inactive",
                "blocked"
              ]
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Filter by type",
            "schema": {
              "example": "contact",
              "type": "string",
              "enum": [
                "contact"
              ]
            }
          },
          {
            "name": "platform",
            "required": false,
            "in": "query",
            "description": "Filter by platform",
            "schema": {
              "example": "whatsapp",
              "type": "string",
              "enum": [
                "whatsapp",
                "telegram",
                "instagram",
                "facebook",
                "other"
              ]
            }
          },
          {
            "name": "priority",
            "required": false,
            "in": "query",
            "description": "Filter by priority",
            "schema": {
              "minimum": 0,
              "maximum": 10,
              "example": 5,
              "type": "number"
            }
          },
          {
            "name": "isGroup",
            "required": false,
            "in": "query",
            "description": "Filter by group contacts",
            "schema": {
              "example": false,
              "type": "boolean"
            }
          },
          {
            "name": "tags",
            "required": false,
            "in": "query",
            "description": "Filter by tags (comma-separated)",
            "schema": {
              "example": "vip,premium",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customers retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CustomerResponseDto"
                      }
                    },
                    "total": {
                      "type": "number"
                    },
                    "page": {
                      "type": "number"
                    },
                    "limit": {
                      "type": "number"
                    },
                    "totalPages": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all customers with pagination and filtering",
        "tags": [
          "Customers"
        ]
      }
    },
    "/customers/find": {
      "get": {
        "operationId": "CustomerController_findCustomerById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "query",
            "description": "Customer ID",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Customer not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Find customer by ID",
        "tags": [
          "Customers"
        ]
      }
    },
    "/customers/update": {
      "patch": {
        "operationId": "CustomerController_updateCustomer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerByIdDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Customer updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "404": {
            "description": "Customer not found"
          },
          "409": {
            "description": "Customer with this platformId, email, CPF, or CNPJ already exists"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update customer by ID with tags",
        "tags": [
          "Customers"
        ]
      }
    },
    "/customers/delete": {
      "delete": {
        "operationId": "CustomerController_deleteCustomer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteCustomerDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Customer deleted successfully"
          },
          "404": {
            "description": "Customer not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete customer by ID",
        "tags": [
          "Customers"
        ]
      }
    },
    "/customers/tags": {
      "get": {
        "operationId": "CustomerController_getAllTags",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all unique customer tags",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "example": [
                    "vip",
                    "premium",
                    "new-customer"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all unique customer tags",
        "tags": [
          "Customers"
        ]
      }
    },
...
    "/dashboard/analytics/cards": {
      "get": {
        "operationId": "AnalyticsController_getCardsMetrics",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Cards metrics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardsMetricsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get dashboard cards metrics",
        "tags": [
          "Dashboard Analytics"
        ]
      }
    },
    "/dashboard/analytics/operators": {
      "get": {
        "operationId": "AnalyticsController_getOperators",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "default": 10,
              "example": 10,
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term for user name, email, or contact",
            "schema": {
              "example": "john",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operators retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get operators with pagination and filtering",
        "tags": [
          "Dashboard Analytics"
        ]
      }
    },
    "/dashboard/analytics/services": {
      "get": {
        "operationId": "AnalyticsController_getActiveServices",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "default": 10,
              "example": 10,
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term for protocol, customer name, or customer number",
            "schema": {
              "example": "john",
              "type": "string"
            }
          },
          {
            "name": "direction",
            "required": false,
            "in": "query",
            "description": "Filter by direction (inbound/outbound)",
            "schema": {
              "enum": [
                "inbound",
                "outbound"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Active services retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveServicesResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get active services with pagination and filtering",
        "tags": [
          "Dashboard Analytics"
        ]
      }
    }
  },
  "info": {
    "title": "CCS Backend API",
    "description": "Multi platform CRM API for customer service",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "Typebot-API-Key": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Typebot API Key (Bearer token)"
      }
    },
    "schemas": {
...
      "CreateCustomerDto": {
        "type": "object",
        "properties": {
          "platformId": {
            "type": "string",
            "description": "Platform-specific ID (WhatsApp phone, Telegram user_id, etc.)",
            "example": "5511999999999",
            "minLength": 1,
            "maxLength": 100
          },
          "pushName": {
            "type": "string",
            "description": "Display name from platform",
            "example": "John Doe",
            "maxLength": 100
          },
          "name": {
            "type": "string",
            "description": "Customer name",
            "example": "John Doe",
            "maxLength": 100
          },
          "profilePicUrl": {
            "type": "string",
            "description": "Profile picture URL",
            "example": "https://example.com/profile.jpg",
            "maxLength": 500
          },
          "contact": {
            "type": "string",
            "description": "Platform phone number",
            "example": "+5511999999999",
            "maxLength": 20
          },
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "john.doe@example.com"
          },
          "cpf": {
            "type": "string",
            "description": "CPF (Brazilian individual tax ID)",
            "example": "12345678901",
            "maxLength": 14
          },
          "cnpj": {
            "type": "string",
            "description": "CNPJ (Brazilian company tax ID)",
            "example": "12345678000195",
            "maxLength": 18
          },
          "donorCode": {
            "type": "string",
            "description": "Donor code",
            "example": "DONOR123",
            "maxLength": 100
          },
          "priority": {
            "type": "number",
            "description": "Priority level for customer service",
            "example": 5,
            "minimum": 0,
            "maximum": 10,
            "default": 0
          },
          "isGroup": {
            "type": "boolean",
            "description": "Whether this is a group contact",
            "example": false,
            "default": false
          },
          "skipBot": {
            "type": "boolean",
            "description": "Whether to skip bot interaction for this customer",
            "example": false,
            "default": false
          },
          "type": {
            "type": "string",
            "description": "Customer type",
            "enum": [
              "contact"
            ],
            "example": "contact",
            "default": "contact"
          },
          "status": {
            "type": "string",
            "description": "Customer status",
            "enum": [
              "active",
              "inactive",
              "blocked"
            ],
            "example": "active",
            "default": "active"
          },
          "platform": {
            "type": "string",
            "description": "Platform type",
            "enum": [
              "whatsapp",
              "telegram",
              "instagram",
              "facebook",
              "other"
            ],
            "example": "whatsapp",
            "default": "whatsapp"
          },
          "observations": {
            "type": "string",
            "description": "Internal notes/observations",
            "example": "VIP customer, prefers morning calls",
            "maxLength": 1000
          },
          "tags": {
            "description": "Customer tags",
            "example": [
              "vip",
              "premium"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "platformId"
        ]
      },
      "CustomerResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Customer unique identifier",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "platformId": {
            "type": "string",
            "description": "Platform-specific ID",
            "example": "5511999999999"
          },
          "pushName": {
            "type": "string",
            "description": "Display name from platform",
            "example": "John Doe"
          },
          "name": {
            "type": "string",
            "description": "Customer name",
            "example": "John Doe"
          },
          "profilePicUrl": {
            "type": "string",
            "description": "Profile picture URL",
            "example": "https://example.com/profile.jpg"
          },
          "contact": {
            "type": "string",
            "description": "Platform phone number",
            "example": "+5511999999999"
          },
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "john.doe@example.com"
          },
          "cpf": {
            "type": "string",
            "description": "CPF (Brazilian individual tax ID)",
            "example": "12345678901"
          },
          "cnpj": {
            "type": "string",
            "description": "CNPJ (Brazilian company tax ID)",
            "example": "12345678000195"
          },
          "donorCode": {
            "type": "string",
            "description": "Donor code",
            "example": "DONOR123"
          },
          "priority": {
            "type": "number",
            "description": "Priority level for customer service",
            "example": 5
          },
          "isGroup": {
            "type": "boolean",
            "description": "Whether this is a group contact",
            "example": false
          },
          "type": {
            "type": "string",
            "description": "Customer type",
            "enum": [
              "contact"
            ],
            "example": "contact"
          },
          "status": {
            "type": "string",
            "description": "Customer status",
            "enum": [
              "active",
              "inactive",
              "blocked"
            ],
            "example": "active"
          },
          "platform": {
            "type": "string",
            "description": "Platform type",
            "enum": [
              "whatsapp",
              "telegram",
              "instagram",
              "facebook",
              "other"
            ],
            "example": "whatsapp"
          },
          "observations": {
            "type": "string",
            "description": "Internal notes/observations",
            "example": "VIP customer, prefers morning calls"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Customer creation timestamp",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Customer last update timestamp",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether customer is active",
            "example": true
          },
          "isBlocked": {
            "type": "boolean",
            "description": "Whether customer is blocked",
            "example": false
          },
          "isHighPriority": {
            "type": "boolean",
            "description": "Whether customer has high priority",
            "example": true
          },
          "isGroupContact": {
            "type": "boolean",
            "description": "Whether this is a group contact",
            "example": false
          },
          "hasProfilePicture": {
            "type": "boolean",
            "description": "Whether customer has profile picture",
            "example": true
          },
          "displayName": {
            "type": "string",
            "description": "Display name for the customer",
            "example": "John Doe"
          },
          "tags": {
            "type": "array",
            "description": "Customer tags",
            "items": {
              "type": "string"
            },
            "example": [
              "vip",
              "premium"
            ]
          }
        },
        "required": [
          "id",
          "platformId",
          "priority",
          "isGroup",
          "type",
          "status",
          "platform",
          "createdAt",
          "updatedAt",
          "isActive",
          "isBlocked",
          "isHighPriority",
          "isGroupContact",
          "hasProfilePicture",
          "displayName",
          "tags"
        ]
      },
      "UpdateCustomerByIdDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Customer unique identifier",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "platformId": {
            "type": "string",
            "description": "Platform-specific ID (WhatsApp phone, Telegram user_id, etc.)",
            "example": "5511999999999",
            "minLength": 1,
            "maxLength": 100
          },
          "pushName": {
            "type": "string",
            "description": "Display name from platform",
            "example": "John Doe",
            "maxLength": 100
          },
          "name": {
            "type": "string",
            "description": "Customer name",
            "example": "John Doe",
            "maxLength": 100
          },
          "profilePicUrl": {
            "type": "string",
            "description": "Profile picture URL",
            "example": "https://example.com/profile.jpg",
            "maxLength": 500
          },
          "contact": {
            "type": "string",
            "description": "Platform phone number",
            "example": "+5511999999999",
            "maxLength": 20
          },
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "john.doe@example.com"
          },
          "cpf": {
            "type": "string",
            "description": "CPF (Brazilian individual tax ID)",
            "example": "12345678901",
            "maxLength": 14
          },
          "cnpj": {
            "type": "string",
            "description": "CNPJ (Brazilian company tax ID)",
            "example": "12345678000195",
            "maxLength": 18
          },
          "donorCode": {
            "type": "string",
            "description": "Donor code",
            "example": "DONOR123",
            "maxLength": 100
          },
          "priority": {
            "type": "number",
            "description": "Priority level for customer service",
            "example": 5,
            "minimum": 0,
            "maximum": 10
          },
          "isGroup": {
            "type": "boolean",
            "description": "Whether this is a group contact",
            "example": false
          },
          "type": {
            "type": "string",
            "description": "Customer type",
            "enum": [
              "contact"
            ],
            "example": "contact"
          },
          "status": {
            "type": "string",
            "description": "Customer status",
            "enum": [
              "active",
              "inactive",
              "blocked"
            ],
            "example": "active"
          },
          "platform": {
            "type": "string",
            "description": "Platform type",
            "enum": [
              "whatsapp",
              "telegram",
              "instagram",
              "facebook",
              "other"
            ],
            "example": "whatsapp"
          },
          "observations": {
            "type": "string",
            "description": "Internal notes/observations",
            "example": "VIP customer, prefers morning calls",
            "maxLength": 1000
          },
          "tags": {
            "description": "Customer tags",
            "example": [
              "vip",
              "premium"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id"
        ]
      },
      "DeleteCustomerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Customer unique identifier",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        },
        "required": [
          "id"
        ]
      },
...
      "CardsMetricsResponseDto": {
        "type": "object",
        "properties": {
          "customersWaitingInQueue": {
            "type": "number",
            "description": "Number of customers waiting in queue",
            "example": 5
          },
          "customersInService": {
            "type": "number",
            "description": "Number of customers in service",
            "example": 12
          },
          "finishedServices": {
            "type": "number",
            "description": "Number of finished services today",
            "example": 45
          },
          "receivedMessages": {
            "type": "number",
            "description": "Number of received messages today",
            "example": 234
          },
          "sentMessages": {
            "type": "number",
            "description": "Number of sent messages today",
            "example": 198
          },
          "readMessages": {
            "type": "number",
            "description": "Number of read messages today",
            "example": 156
          },
          "total": {
            "type": "number",
            "description": "Total messages (received + sent) today",
            "example": 432
          }
        },
        "required": [
          "customersWaitingInQueue",
          "customersInService",
          "finishedServices",
          "receivedMessages",
          "sentMessages",
          "readMessages",
          "total"
        ]
      },
...
    }
  }
}